refactor: change state structure
