+++ src/app/redux/actions/user.action.ts
@@ -1,5 +1,9 @@
 import { createAction, props } from '@ngrx/store';
-import { User, UserStateInterface } from 'src/app/shared/models/user.model';
+import {
+  User,
+  UserBooking,
+  UserStateInterface,
+} from 'src/app/shared/models/user.model';
 
 export const registerUser = createAction(
   '[User] Register user on server',
@@ -9,4 +13,14 @@ export const registerUser = createAction(
 export const addUserToState = createAction(
   '[User] Add user to store',
   props<{ user: UserStateInterface }>()
+);
+
+export const addBookingToServer = createAction(
+  '[Booking] Add booking to the server',
+  props<{ booking: UserBooking }>
+);
+
+export const addBookingsToState = createAction(
+  '[Booking] Add booking to state',
+  props<{ booking: UserBooking }>()
 );
