+++ src/app/redux/selectors/user.selectors.ts
@@ -7,3 +7,18 @@ export const selectUserDate = createSelector(
   selectUser,
   userState => userState.userDate
 );
+
+// const selectBookings = createFeatureSelector<UserState>('bookings');
+
+export const selectAllBookings = createSelector(
+  selectUser,
+  userState => userState.bookings
+);
+
+export const selectUnpaidBookings = createSelector(selectUser, userState =>
+  userState.bookings.filter(item => item.paid === false)
+);
+
+export const selectPaidBookings = createSelector(selectUser, userState =>
+  userState.bookings.filter(item => item.paid === true)
+);
